+++ yt/server/cypress_server/node_proxy_detail.h
@@ -316,7 +316,7 @@ private:
     }
 
 DECLARE_SCALAR_TYPE(String, Stroka)
-DECLARE_SCALAR_TYPE(Integer, i64)
+DECLARE_SCALAR_TYPE(Int64, i64)
 DECLARE_SCALAR_TYPE(Double, double)
 
 #undef DECLARE_SCALAR_TYPE
