+++ yt/ytlib/query_client/cg_types.h
@@ -177,7 +177,7 @@ public:
 
     enum Fields
     {
-        Integer,
+        Int64,
         Double,
         String
     };
